test(engine): convert LocalBackedJdbcHashLabel spec to JUnit, drop redundant case
Rewrite the kotest StringSpec as a JUnit 5 test and drop the write
round-trip case, which duplicates ByteArrayLabelCompatibilityTest. Keep
only the indexName=null -> __default__ scan routing, which is unique to
LocalBackedJdbcHashLabel.
